backend/pyproject.toml
[project]
name = "dnd-rulebook-llm-backend"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn",
    "langchain",
    "langchain-community",
    "langchain-huggingface",
    "langchain-openai",
    "langchain-chroma",
    "python-dotenv",
    "llama-cloud-services",
    "sentence-transformers",
]